code formatted, getAllTask renamed

+++ src/components/Task.js
@@ -9,7 +9,7 @@ import useTaskStore from "../store/taskStore";
 const baseApi = process.env.REACT_APP_BASE_API;
 
 const Task = (task) => {
-  const { refreshTasks: getAllTask } = useTaskStore();
+  const { refreshTasks } = useTaskStore();
 
   const updateTask = async (taskId, completed) => {
     customLogger('Updating Task');
@@ -23,7 +23,7 @@ const Task = (task) => {
           Authorization: `Bearer ${token}`,
         },
       }).then((res) => {
-        getAllTask().then(() => {
+        refreshTasks().then(() => {
           toast.success('Task Updated Successfully');
         });
       }).catch((err) => {
@@ -37,25 +37,27 @@ const Task = (task) => {
     customLogger('Deleting Task');
     toast.info('Deleting Task');
     const token = getUserToken()
-    await axios.delete(`${baseApi}/tasks/${taskId}`, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    }).then((res) => {
-      getAllTask().then(() => {
-        toast.success('Task Deleted Successfully');
+    if (token) {
+      await axios.delete(`${baseApi}/tasks/${taskId}`, {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      }).then((res) => {
+        refreshTasks().then(() => {
+          toast.success('Task Deleted Successfully');
+        });
+      }).catch((err) => {
+        httpErrorLogger(err);
+        toast.error('Error Deleting Task');
       });
-    }).catch((err) => {
-      httpErrorLogger(err);
-      toast.error('Error Deleting Task');
-    });
+    }
   };
 
   return (
     <TaskStyle isDone={task.task.completed}>
       <h4>{task.task.completed ? <span style={{ textDecoration: 'line-through' }}>{task.task.description}</span> : task.task.description}</h4>
       <ButtonRow>
-        <Button isDeleteButton={false} onClick={() => updateTask(task.task._id, task.task.completed)}>{task.task.completed? 'Undo': 'Done'}</Button>
+        <Button isDeleteButton={false} onClick={() => updateTask(task.task._id, task.task.completed)}>{task.task.completed ? 'Undo' : 'Done'}</Button>
         <Button isDeleteButton={true} onClick={() => deleteTask(task.task._id)}>Delete</Button>
       </ButtonRow>
       <ToastContainer theme="dark" />
